google marker updates 50% complete

+++ InstaTask/app/src/main/java/com/example/instatask/database/datamodel/Task.kt
@@ -36,8 +36,8 @@ data class Task (
     val address:String? = null,
 
     @ColumnInfo(name = "lat")
-    val lat:Float? = null,
+    val lat:Double? = null,
 
     @ColumnInfo(name = "long")
-    val lng:Float? = null
+    val lng:Double? = null
 )
